Fixing SVG lines that weren't rendering in Butter.
Some exported com.sun.star.drawing.LineShape instances wouldn't render
in Butter, though they would render in the exported project. It turns
out that these ones didn't just define the line to be drawn, but also
applied a mask filled with a linear gradient, for some reason. The
"gradient" is actually one solid colour and the mask seems to have the
wrong coordinates. The lines also failed to render when used as an
embedded SVG in a jsfiddle preview. However, they work correctly in
the exported document.

I'm not sure exactly what's going on, but removing the mask element
fixes it and none of the examples we're working with actually
require the mask for any purpose.